Rename ProjectList file to match operation name
